Refactor EVM functions and weights
